% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENQviewpkg-package.R
\docType{package}
\name{ENQviewpkg-package}
\alias{ENQviewpkg}
\alias{ENQviewpkg-package}
\title{ENQviewpkg: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: kazuo fujimoto \email{kazuo.fujimoto2007@gmail.com}

}
\keyword{internal}
