% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_ENQview}
\alias{run_ENQview}
\title{ENQview_lite アプリを起動する関数}
\usage{
run_ENQview(df = NULL)
}
\arguments{
\item{df}{作業中のデータフレーム（オプショナル）。渡された場合はそのデータを優先して使用します。}
}
\description{
ENQview_lite アプリを起動する関数
}
